# more type-stable type-inference